+++ LegendWeathers/Plugin.cs
@@ -29,7 +29,7 @@ namespace LegendWeathers
     {
         const string GUID = "zigzag.legendweathers";
         const string NAME = "LegendWeathers";
-        const string VERSION = "2.1.0";
+        const string VERSION = "2.1.1";
 
         public static Plugin instance;
         public static ManualLogSource logger;